Create Dice Rolling Simulator with Art Display
This is a dice rolling simulator that not only generates random dice values but also displays corresponding ASCII art of dice faces. The code features a collection of beautifully crafted ASCII art patterns representing dice faces from 1 to 6. When executed, the user is prompted to input the number of sides on the virtual dice, ranging from 1 to a maximum value of their choice. Upon each roll, the program presents the rolled dice value along with an artistic rendering of the dice face. The user is then given the option to roll the dice again or exit the program.